% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cobweb}
\alias{cobweb}
\title{cobweb}
\usage{
cobweb(
  city,
  span = 2000,
  nroad = 1000,
  ncl = 1,
  filename = "map.png",
  verbose = FALSE,
  url = "http://0.0.0.0:5000/"
)
}
\arguments{
\item{city}{city name with postal code if possible and country name in English}

\item{span}{max road distance in km}

\item{nroad}{number of roads}

\item{ncl}{number of cores}

\item{filename}{file name of the map produced}

\item{verbose}{output messages}

\item{url}{OSRM server URL}
}
\value{
Nothing is returned, a map is exported.
}
\description{
Build a city cobweb
}
\examples{
\dontrun{
cobweb("Marseille, 13000 France")
}
}
