% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_routes}
\alias{get_routes}
\title{Get Time and Distance Between Points}
\usage{
get_routes(x, src, dst, ncl = 1)
}
\arguments{
\item{x}{a data.frame}

\item{src}{names of the variables containing source coordinates (WGS84)}

\item{dst}{names of the variables containing destination coordinates (WGS84)}

\item{ncl}{number of cores}
}
\value{
a matrix containing duration (minutes) and distance (km) is returned.
}
\description{
Use this function to get time and distance between points
}
\examples{
library(sf)

p1 <- st_as_sf(data.frame(x = 2, y = 45), coords = c("x", "y"), crs = 4326)
p2 <- st_as_sf(data.frame(x = 10, y = 45), coords = c("x", "y"), crs = 4326)

z1 <- st_buffer(p1, 50000)
z2 <- st_buffer(p2, 50000)

pts1 <- st_sample(z1, 100)
pts2 <- st_sample(z2, 100)

tab <- data.frame(st_coordinates(pts1), st_coordinates(pts2))
names(tab) <- c("Xs", "Ys", "Xd", "Yd")
res <- get_routes(x = tab, src = c("Xs", "Ys"), dst = c("Xd", "Yd"), ncl = 8)
cbind(tab, res)
}
