% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cobweb}
\alias{cobweb}
\title{cobweb}
\usage{
cobweb(city, filename = "map.png")
}
\arguments{
\item{city}{city name with postal code if possible and country name in english}

\item{filename}{filename of the map produced}
}
\value{
a map is exported
}
\description{
Build a city cobweb
}
\examples{
\dontrun{
cobweb("Marseille, 13000 France")
}
}
